# packages/contracts/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateProfileUpdate": {
        "additionalProperties": false,
        "properties": {
          "default_interview_mode": {
            "enum": [
              "COACH",
              "SIMULATION"
            ],
            "title": "Default Interview Mode",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "interview_level": {
            "enum": [
              "INTERN",
              "NEW_GRAD",
              "EARLY_CAREER"
            ],
            "title": "Interview Level",
            "type": "string"
          },
          "preferred_language": {
            "enum": [
              "cpp",
              "java",
              "python"
            ],
            "title": "Preferred Language",
            "type": "string"
          },
          "target_role": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Role"
          },
          "timezone": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "required": [
          "preferred_language",
          "default_interview_mode",
          "interview_level"
        ],
        "title": "CandidateProfileUpdate",
        "type": "object"
      },
...
      "CurrentUserProfileResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_interview_mode": {
            "enum": [
              "COACH",
              "SIMULATION"
            ],
            "title": "Default Interview Mode",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "interview_level": {
            "enum": [
              "INTERN",
              "NEW_GRAD",
              "EARLY_CAREER"
            ],
            "title": "Interview Level",
            "type": "string"
          },
          "preferred_language": {
            "enum": [
              "cpp",
              "java",
              "python"
            ],
            "title": "Preferred Language",
            "type": "string"
          },
          "profile_version": {
            "title": "Profile Version",
            "type": "integer"
          },
          "target_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Role"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "display_name",
          "preferred_language",
          "default_interview_mode",
          "interview_level",
          "target_role",
          "timezone",
          "profile_version",
          "created_at",
          "updated_at"
        ],
        "title": "CurrentUserProfileResponse",
        "type": "object"
      },
      "CurrentUserResponse": {
        "properties": {
          "account_status": {
            "title": "Account Status",
            "type": "string"
          },
          "onboarding_required": {
            "title": "Onboarding Required",
            "type": "boolean"
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurrentUserProfileResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "account_status",
          "onboarding_required",
          "profile"
        ],
        "title": "CurrentUserResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "CounterQ application API.",
    "title": "CounterQ API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/me": {
      "get": {
        "operationId": "current_user_api_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Current User",
        "tags": [
          "current-user"
        ]
      }
    },
    "/api/me/profile": {
      "put": {
        "operationId": "update_current_user_profile_api_me_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Current User Profile",
        "tags": [
          "current-user"
        ]
      }
    },
...
  }
}